requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[standard]
pytest
pytest-asyncio
requests
pycryptodome
setuptools